+++ src/components/game/baseObjects/cell/TowerCell.js
@@ -1,17 +1,17 @@
 import Cell from "./Cell";
-import { CELL_TYPE } from "../../../Constants";
-import { drawRect, outlineRect } from "../../../graphics/Pipeline";
+import { CELL_TYPE } from "../../Constants";
+import { drawRect, outlineRect } from "../../graphics/Pipeline";
 
 import {
     EMPTY_COLOR,
     LIGHT_AMPLITUDE,
-} from "../../../graphics/Theme";
+} from "../../graphics/Theme";
 import {
     interpolateColor,
-} from "../../../graphics/utils/Colors";
+} from "../../graphics/utils/Colors";
 import {
     linInt,
-} from "../../../graphics/utils/Functions";
+} from "../../graphics/utils/Functions";
 
 
 class TowerCell extends Cell {
